Show a confirmation dialog before first darkening the screen.
The dialog will be shown each time the user tries to darken the
screen until the user "confirms" it. After that, the dialog will
not be shown again.

This is not synced between chromebooks.

This change also improves styling of accelerator dialogs by
left-justifying the body text and fixing the size to the standard
dialog width.

tries to darken the screen in ChromeVox. Improves styling of
accelerator dialogs.

Bug: 1103799
AX-Relnotes: A confirmation dialog is shown the first time the user
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2770182
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864422}

+++ chrome/browser/resources/chromeos/accessibility/chromevox/background/command_handler.js
@@ -247,11 +247,25 @@ CommandHandler.onCommand = function(command) {
     case 'toggleDarkScreen':
       const oldState = sessionStorage.getItem('darkScreen');
       const newState = (oldState === 'true') ? false : true;
-      sessionStorage.setItem('darkScreen', (newState) ? 'true' : 'false');
-      chrome.accessibilityPrivate.darkenScreen(newState);
-      new Output()
-          .format((newState) ? '@darken_screen' : '@undarken_screen')
-          .go();
+      if (newState && localStorage['acceptDarkenScreen'] !== 'true') {
+        // If this is the first time, show a confirmation dialog.
+        chrome.accessibilityPrivate.showConfirmationDialog(
+            Msgs.getMsg('darken_screen'),
+            Msgs.getMsg('darken_screen_description'), (confirmed) => {
+              if (confirmed) {
+                sessionStorage.setItem('darkScreen', 'true');
+                localStorage['acceptDarkenScreen'] = true;
+                chrome.accessibilityPrivate.darkenScreen(true);
+                new Output().format('@darken_screen').go();
+              }
+            });
+      } else {
+        sessionStorage.setItem('darkScreen', (newState) ? 'true' : 'false');
+        chrome.accessibilityPrivate.darkenScreen(newState);
+        new Output()
+            .format((newState) ? '@darken_screen' : '@undarken_screen')
+            .go();
+      }
       return false;
     case 'toggleSpeechOnOrOff':
       const state = ChromeVox.tts.toggleSpeechOnOrOff();
